Refactor code

+++ app/Repository/KurbanHewanRepository.php
@@ -8,30 +8,35 @@ use App\Repository\Interfaces\KurbanHewanRepositoryInterface;
 
 class KurbanHewanRepository implements KurbanHewanRepositoryInterface
 {
-    public function allData()
-    {
-        return Model::UserMasjid()->latest()->paginate(50);
-    }
+  public function allData()
+  {
+    return Model::UserMasjid()->latest()->paginate(50);
+  }
 
-    public function storeData($request): void
-    {
-        $requestData = $request->validated();
-        Model::create($requestData);
-    }
+  public function storeData($request): void
+  {
+    $requestData = $request->validated();
+    Model::create($requestData);
+  }
 
-    public function findData($data)
-    {
-        return $data;
-    }
+  public function findData($data)
+  {
+    return $data;
+  }
 
-    public function updateData($request, $kurban)
-    {
-        $requestData = $request->validated();
-        $kurban->update($requestData);
-    }
+  public function updateData($request, $kurban)
+  {
+    $requestData = $request->validated();
+    $kurban->update($requestData);
+  }
 
-    public function destroyData($data)
-    {
-        $data->delete();
+  public function destroyData($data)
+  {
+    if ($data->kurbanPeserta->count() == 0) {
+      $data->delete();
+      return true;
+    } else {
+      return false;
     }
+  }
 }
